refactor(language): split language tasks into separate files
